requirements.txt
matplotlib==3.7.3
numpy==1.24.4
onnxruntime==1.17.3
opencv_python==4.8.1.78
pyrealsense2==2.54.2.5684
pyzmq==25.1.1
scipy==1.13.0
skimage==0.0
smbus==1.1.post2
torch==2.1.0
torchvision==0.16.0
RPi.GPIO==0.7.1